requirements.txt
confluent_kafka
python-dotenv
praw
faker
findspark
psycopg2-binary